Several bug fixes revealed during unit testing.
The most significant change was that abbreviation nodes in ctrees
are named according to a hash of the ctree, not its hashroot.
This is necessary since many different ctrees can have the same hashroot,
and the abbreviation should be clearly identifying which of these is intended.
The hashroot is still saved as the abbreviation node so that the hashroot
can be computed without expanding abbreviation nodes. Abbreviation nodes
are stored in files (named by the hashval) in the data directory.
The variable holding the location of the data directory was moved from
ctre to config.

Also, ctree_reduce_to_min_support needed to be rewritten to traverse properly through
leaf nodes so that ultimately the nehlist will be abstracted. The way it was written
before, all leaf nodes not at level 0 (in practice, all of them) would keep the full nehlist,
which is unrealistic. The Coq version (qeditastheory) still corresponds to the old
version which does not minimize leaf nodes enough.

Other changes refine the test for when a ctree supports a transaction.
Publications now need to be new (checked by ensuring the leaf of the ctree
at the publication address is empty). When an ownership asset is in the output,
we now first check whether a corresponding asset is in the input (so that ownership
is being transfered).

There were also some minor bug fixes.