changed staking so that three kinds of currency assets stake differently:
. unlocked (None obligation) - matures after 512 blocks and then ages quadratically
 for a few months to reach a maximum age (exception: those from the initial distribution
 have birthday 0 and start at maximum age)
. reward (obligation unlocking at later block n; marked as reward) - matures after 512 blocks
 and then ages like the unlocked case until 32 blocks before it will be unlocked;
 once it reaches the almost-unlocked stage it cannot be used to stake anymore
. locked (obligation unlocking at later block n; not reward) - matures after 8 blocks
 and then reaches maximum age until 32 blocks before it will be unlocked;
 once it reaches the almost-unlocked stage it cannot be used to stake anymore

One reason for handling the "locked" case this way is in case there
are very few stakers during the initial 512 blocks before the first
stake rewards become eligible to stake.  With very few stakers and the
old rules, there was a good chance of reaching a deadlock during those
initial 512 blocks. With the new rules, someone can spend some of his
or her initial distribution to several locked addresses which will be
able to stake within 8 blocks.