Several changes resulting from test runs of qeditasd with testnet staking.
The largest changes are as follows:

There are now files storing both headers and validatedheaders (as hashes)
so that headers which were shown to correspond to valid blocks in the past
need not re-validate the blocks on initialization.

During the initialization phase, the code for reading headers
tries to recognize if qednetd is not ready and keep retrying once
a second until it is ready.

blocktreenodes were changed in two ways: 1. The node now stores
both the previous target info and the current target info. The
previous target info is needed to check if the new blocks are valid
successor blocks. 2. The validation ref is now of type validationstatus
which may be Waiting(float) (indicating the time from which the
block header has waited for a corresponding block delta to be validated),
Invalid or Valid.

After a block is validated, bestnode may be updated to be the corresponding node.
The usual case will be that the header is already the bestnode. However,
if the block took more than 60s to be obtained and validated, bestnode
may have fallen back to a previous validated node.

Some debugging print statements have been removed.